AuraFlow-Genisis-Backend/requirements.txt
pinecone-client==3.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
ollama==0.1.5
tqdm==4.66.1
colorama==0.4.6
pydantic==2.5.2
langchain-core==0.1.18
langchain-pinecone==0.1.1
